Assume that when DateTime.Kind is Unspecified, the date is actually Local. Fixes #3983.
This is required because NSDate doesn't have the concept of an unspecified time zone,
so at some point we have to decide what an Unspecified kind is. The implicit conversion
operators between NSDate and DateTime already assumes that Unspecified means Local, so
just update the UI to show the DateTime values correctly.

Note that ToLocalTiem does not do this: it assumes the input is Utc if it's
unspecified, which is why we have to do it explicitly before calling ToLocalTime.